F7: replace std with OU sigma
Used to be computing the standard deviation of the binned raster and
converting that to units of firing rate. The more principled method is
to actually estimate the parameters of the Ornstein-Uhlenbeck process
(by fitting a first-order autoregressive process, subtracting it, and
calculating σ to be the standard deviation of the residuals corrected
for timestep).

Also actually compute the square root law in the figure instead of using
the manually input value from before. :)